feat: add @xm-ngx/components/xm-ribbon
