+++ packages/components/src/xm-ribbon/index.ts
@@ -1,0 +1,1 @@
+export { XmRibbonModule } from './xm-ribbon.module';
