+++ packages/components/src/xm-ribbon/xm-ribbon.component.html
@@ -1,4 +1,4 @@
-<ng-container *ngIf="(show$ | async) && serverEnv">
+<ng-container *ngIf="!config?.hidden && serverEnv">
   <div class="xm-ribbon">
     <div class="xm-ribbon-text">server: {{serverEnv}}</div>
     <div class="xm-ribbon-text">release: {{environment.release}}</div>
