+++ packages/components/src/xm-ribbon/xm-ribbon.component.scss
@@ -1,4 +1,4 @@
-@import 'theme';
+@import 'src/styles/theme';
 
 $xm-ribbon-opacity: 0.4 !default;
 $xm-ribbon-z-index: 9999 !default;
