+++ packages/components/src/xm-ribbon/xm-ribbon.component.spec.ts
