+++ packages/components/src/xm-ribbon/xm-ribbon.module.ts
@@ -1,0 +1,12 @@
+import { CommonModule } from '@angular/common';
+import { NgModule, Type } from '@angular/core';
+import { XmRibbonComponent } from './xm-ribbon.component';
+
+@NgModule({
+    declarations: [XmRibbonComponent],
+    exports: [XmRibbonComponent],
+    imports: [CommonModule],
+})
+export class XmRibbonModule {
+    public entry: Type<XmRibbonComponent> = XmRibbonComponent;
+}
