+++ src/app/layouts/main/main.component.html
@@ -1,5 +1,4 @@
 <xm-page-ribbon></xm-page-ribbon>
-<xm-ribbon></xm-ribbon>
 <xm-alerts></xm-alerts>
 <xm-alert-error></xm-alert-error>
 <xm-password-needed></xm-password-needed>
