--- src/app/modules/xm-ribbon/xm-ribbon.module.ts
@@ -1,11 +1,0 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {XmRibbonComponent} from './xm-ribbon.component';
-
-@NgModule({
-    declarations: [XmRibbonComponent],
-    exports: [XmRibbonComponent],
-    imports: [CommonModule],
-})
-export class XmRibbonModule {
-}
